Refactor settings import, update Docker test steps, and enhance settings management.
