--- .github/workflows/ci.yml
@@ -1,135 +1,0 @@
-name: Continuous Integration
-
-on:
-  push:
-    branches: [ main ]
-    paths-ignore:
-      - '**.md'
-      - 'docs/**'
-  pull_request:
-    paths-ignore:
-      - '**.md'
-      - 'docs/**'
-
-jobs:
-  quality:
-    name: Code Quality
-    runs-on: ubuntu-24.04
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-          cache: 'pip'
-
-      - name: Install Poetry
-        run: pipx install poetry
-
-      - name: Configure Poetry
-        run: |
-          poetry config virtualenvs.in-project true
-          poetry config installer.parallel true
-
-      - name: Cache Poetry Dependencies
-        uses: actions/cache@v4
-        with:
-          path: |
-            ~/.cache/pypoetry
-            .venv
-          key: ${{ runner.os }}-poetry-${{ hashFiles('poetry.lock') }}
-          restore-keys: |
-            ${{ runner.os }}-poetry-
-
-      - name: Install Dependencies
-        run: poetry install --no-interaction
-
-      - name: Run Format Check
-        run: poetry run invoke check-format
-
-      - name: Run Linting
-        run: poetry run invoke lint
-
-      - name: Run Type Check
-        run: poetry run mypy backend_core
-
-  test:
-    name: Tests
-    runs-on: ubuntu-24.04
-    services:
-      postgres:
-        image: postgres:16
-        env:
-          POSTGRES_USER: postgres
-          POSTGRES_PASSWORD: postgres
-          POSTGRES_DB: postgres
-        ports:
-          - 5432:5432
-        options: >-
-          --health-cmd pg_isready
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-          cache: 'pip'
-
-      - name: Install Poetry
-        run: pipx install poetry
-
-      - name: Configure Poetry
-        run: |
-          poetry config virtualenvs.in-project true
-          poetry config installer.parallel true
-
-      - name: Cache Poetry Dependencies
-        uses: actions/cache@v4
-        with:
-          path: |
-            ~/.cache/pypoetry
-            .venv
-          key: ${{ runner.os }}-poetry-${{ hashFiles('poetry.lock') }}
-          restore-keys: |
-            ${{ runner.os }}-poetry-
-
-      - name: Install Dependencies
-        run: poetry install --no-interaction
-
-      - name: Run Tests
-        run: poetry run invoke test
-        env:
-          DATABASE_URL: postgresql://postgres:postgres@localhost:5432/postgres
-          SECRET_KEY: test_secret_key
-          FIRST_SUPERUSER_EMAIL: admin@example.com
-          FIRST_SUPERUSER_PASSWORD: admin
-
-      - name: Upload Coverage
-        uses: codecov/codecov-action@v4
-        with:
-          token: ${{ secrets.CODECOV_TOKEN }}
-          files: ./coverage.xml
-          fail_ci_if_error: true
-
-  security:
-    name: Security Checks
-    runs-on: ubuntu-24.04
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Run Trivy vulnerability scanner
-        uses: aquasecurity/trivy-action@master
-        with:
-          scan-type: 'fs'
-          scan-ref: '.'
-          format: 'table'
-          exit-code: '1'
-          ignore-unfixed: true
-          severity: 'CRITICAL,HIGH'
-
-      - name: Run dependency review
-        uses: actions/dependency-review-action@v3
