--- .github/workflows/test.yml
@@ -1,29 +1,0 @@
-name: Test
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v4
-    
-    - name: Set up Docker Buildx
-      uses: docker/setup-buildx-action@v3
-      
-    - name: Run tests in Docker
-      run: |
-        docker compose -f docker-compose.test.yml up --build --abort-on-container-exit --exit-code-from test
-      env:
-        DOCKER_BUILDKIT: 1
-        
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v3
-      with:
-        file: ./coverage.xml
-        fail_ci_if_error: true
