+++ backend_core/core/settings.py
@@ -1,21 +1,23 @@
 """Application settings management."""
 
 from functools import lru_cache
-from typing import List
+from pathlib import Path
+from typing import List, Union
 
-from pydantic import AnyHttpUrl, Field, PostgresDsn
+from pydantic import AnyHttpUrl, Field, PostgresDsn, field_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 
 class Settings(BaseSettings):
     """Application settings managed as a singleton."""
 
-    model_config = SettingsConfigDict(env_file=".env", case_sensitive=True)
-
     # API configuration
     API_V1_STR: str = "/api/v1"
     PROJECT_NAME: str = "Backend Core"
     VERSION: str = "0.1.0"
+
+    # Environment
+    TARGET: str = "development"
 
     # Security
     SECRET_KEY: str = Field(..., alias="SECRET_KEY")
@@ -25,17 +27,52 @@ class Settings(BaseSettings):
     # CORS
     BACKEND_CORS_ORIGINS: List[AnyHttpUrl] = []
 
-    # Database
-    DATABASE_URL: PostgresDsn = Field(..., alias="DATABASE_URL")
+    @field_validator("BACKEND_CORS_ORIGINS", mode="before")
+    def assemble_cors_origins(cls, v: Union[str, List[str]]) -> Union[List[str], str]:
+        """Validate CORS origins."""
+        if isinstance(v, str) and not v.startswith("["):
+            return [i.strip() for i in v.split(",")]
+        elif isinstance(v, (list, str)):
+            return v
+        raise ValueError(v)
 
-    # Target environment
-    TARGET: str = "development"
+    # PostgreSQL Database
+    POSTGRES_SERVER: str = Field(..., alias="POSTGRES_SERVER")
+    POSTGRES_USER: str = Field(..., alias="POSTGRES_USER")
+    POSTGRES_PASSWORD: str = Field(..., alias="POSTGRES_PASSWORD")
+    POSTGRES_DB: str = Field(..., alias="POSTGRES_DB")
+    POSTGRES_PORT: str = Field(..., alias="POSTGRES_PORT")
+
+    @property
+    def DATABASE_URL(self) -> str:
+        """Get database URL."""
+        return str(
+            PostgresDsn.build(
+                scheme="postgresql",
+                username=self.POSTGRES_USER,
+                password=self.POSTGRES_PASSWORD,
+                host=self.POSTGRES_SERVER,
+                port=int(self.POSTGRES_PORT),
+                path=self.POSTGRES_DB,
+            )
+        )
+
+    # First superuser
+    FIRST_SUPERUSER: str = Field(..., alias="FIRST_SUPERUSER")
+    FIRST_SUPERUSER_PASSWORD: str = Field(..., alias="FIRST_SUPERUSER_PASSWORD")
+
+    model_config = SettingsConfigDict(
+        case_sensitive=True,
+        env_file=".env.test" if Path(".env.test").exists() else ".env",
+        extra="ignore",
+        env_file_encoding="utf-8",
+    )
 
 
-@lru_cache
+@lru_cache()
 def get_settings() -> Settings:
     """Get cached settings instance."""
-    return Settings()
+    return Settings()  # type: ignore[call-arg]
 
 
 # Create singleton instance
