+++ backend_core/db/session.py
@@ -5,7 +5,7 @@ from typing import Generator
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend_core.core.config import settings
+from backend_core.core.settings import settings
 
 # Create database engine
 engine = create_engine(str(settings.DATABASE_URL), pool_pre_ping=True, pool_size=5, max_overflow=10)
